binary_extractor/requirements.txt
opencv-contrib-python-headless>=4.8.0
scikit-image>=0.21.0
mahotas>=1.4.13
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pytest>=7.4.0
rich>=13.0.0
pyyaml>=6.0 
scipy>=1.10.0
# realesrgan-ncnn-vulkan: install binary from GitHub releases, not pip 